+++ CHANGES.md
@@ -1,3 +1,12 @@
+unreleased
+----------
+
+- Support `# check` parser directive. (@MisterDA, #241)
+- Support `RUN` new `--device` option. (@MisterDA, #241)
+- Support `ADD` new `--unpack` flag. (@MisterDA, #241)
+- Add a `layers` function to estimate the number of Dockerfile layers.
+  (@fiifikcbentil, #239, #240)
+
 v8.3.1 2025-09-15
 -----------------